refactor: move fmtError to shared utils module
Moves fmtError from foreman.ts to a new src/utils.ts and uses it in
repl.ts, worker.ts, workspace.ts, and config.ts to fix [object Object]
error logging in those files.

Closes #324
